// src/entities/token/model/types.ts
export type Token = {
  ammId: string;
  baseMint: string;
  liquidity: number;
  pairName: string;
  quoteMint: string;
  tokenExtensions: object;
  tokenDescription: string;
  tokenImage: string;
  tokenName: string;
  tokenSymbol: string;
  updatedAt: string;
  volume7d: number;
  volume24h: number;
  volume30d: number;
};